Refactor basket service and add new functionality

+++ src/modules/basket/basket.service.ts
@@ -1,8 +1,8 @@
-import { BadRequestException, Inject, Injectable, NotFoundException } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { IsNull, Not, Repository } from "typeorm";
-import { REQUEST } from "@nestjs/core";
 import { Request } from "express";
+import { REQUEST } from "@nestjs/core";
+import { IsNull, Not, Repository } from "typeorm";
+import { InjectRepository } from "@nestjs/typeorm";
+import { BadRequestException, Inject, Injectable, NotFoundException } from "@nestjs/common";
 
 import { BooksService } from "../books/books.service";
 import { BasketEntity } from "./entities/basket.entity";
@@ -32,10 +32,7 @@ export class BasketService {
 		if (basketItem) {
 			basketItem.count += 1;
 		} else {
-			basketItem = this.basketRepository.create({
-				bookId,
-				userId,
-			});
+			basketItem = this.basketRepository.create({ bookId, userId });
 		}
 		await this.basketRepository.save(basketItem);
 		return { message: BasketMessage.AddItemToBasket };
@@ -65,7 +62,7 @@ export class BasketService {
 	async getBasket() {
 		const { id: userId } = this.req.user;
 		const basketItems = await this.basketRepository.find({
-			where: { userId },
+			where: { userId, is_active: true },
 			relations: {
 				discount: true,
 				book: { images: true },
@@ -74,57 +71,65 @@ export class BasketService {
 
 		const books = basketItems.filter((item) => item.bookId);
 		const generalDiscounts = basketItems.find((item) => item?.discount?.id);
+
 		let total_amount = 0;
 		let payment_amount = 0;
 		let total_discount_amount = 0;
 		let bookList = [];
 
 		for (const item of books) {
-			let discount_amount = 0;
-			let discountCode: string = null;
 			const { book, count } = item;
-			total_amount += book.price * count;
+			let discount_amount = 0;
 			let bookPrice = book.price * count;
+
 			if (book.is_active && book.discount > 0) {
-				discount_amount += bookPrice * (book.discount / 100);
-				bookPrice = bookPrice - bookPrice * (book.discount / 100);
+				discount_amount = bookPrice * (book.discount / 100);
+				bookPrice -= discount_amount;
 			}
 
+			total_amount += bookPrice + discount_amount;
 			payment_amount += bookPrice;
 			total_discount_amount += discount_amount;
+
 			bookList.push({
+				bookId: book.id,
 				name: book.name,
 				description: book.introduction,
 				count,
-				image: book.images,
 				price: book.price,
-				total_amount: book.price * count,
+				discountCode: null,
 				discount_amount,
-				payment_amount: book.price * count - discount_amount,
-				discountCode,
+				total_amount: book.price * count,
+				payment_amount: bookPrice,
+				image: book.images,
 			});
 		}
+
 		let generalDiscountDetail = {};
 		if (generalDiscounts?.discount?.active) {
 			const { discount } = generalDiscounts;
-			if (discount?.limit && discount.limit > discount.usage) {
-				let discount_amount = 0;
+
+			const isLimitNotExceeded = discount?.limit === null || discount.limit > discount.usage;
+
+			if (isLimitNotExceeded) {
+				let generalDiscount_amount = 0;
 				if (discount.percent > 0) {
-					discount_amount = payment_amount * (discount.percent / 100);
+					generalDiscount_amount = payment_amount * (discount.percent / 100);
 				} else if (discount.amount > 0) {
-					discount_amount = discount.amount;
+					generalDiscount_amount = discount.amount;
 				}
-				payment_amount =
-					discount_amount > payment_amount ? 0 : payment_amount - discount_amount;
-				total_discount_amount += discount_amount;
+				payment_amount -= generalDiscount_amount;
+				total_discount_amount += generalDiscount_amount;
+
 				generalDiscountDetail = {
 					code: discount.code,
 					percent: discount.percent,
 					amount: discount.amount,
-					discount_amount,
+					discount_amount: generalDiscount_amount,
 				};
 			}
 		}
+
 		return {
 			total_amount,
 			payment_amount,
@@ -132,6 +137,25 @@ export class BasketService {
 			bookList,
 			generalDiscountDetail,
 		};
+	}
+
+	async basketDisable(userId: number) {
+		await this.basketRepository.update({ userId }, { is_active: false });
+		return true;
+	}
+
+	async getBasketDiscount(userId: number) {
+		const basketItems = await this.basketRepository.find({
+			where: { userId, is_active: true },
+			relations: { discount: true },
+		});
+
+		const generalDiscounts = basketItems.find((item) => item?.discount?.id);
+		if (generalDiscounts?.discount) {
+			await this.discountService.useDiscount(generalDiscounts?.discount?.code);
+			return true;
+		}
+		return false;
 	}
 
 	async addDiscount(discountDto: DiscountBasketDto) {
@@ -149,6 +173,7 @@ export class BasketService {
 
 		const userBasketDiscount = await this.basketRepository.findOneBy({
 			userId,
+			is_active: true,
 			discountId: discount.id,
 		});
 
@@ -163,13 +188,16 @@ export class BasketService {
 			},
 		});
 
-		if (generalDiscount?.discount?.code == code) {
-			throw new BadRequestException(BasketMessage.AlreadyDiscount);
-		} else if (generalDiscount?.discount?.code != code) {
-			await this.basketRepository.delete({ discountId: generalDiscount.discountId, userId });
+		if (generalDiscount) {
+			console.log(generalDiscount);
+			if (generalDiscount?.discount?.code == code) {
+				throw new BadRequestException(BasketMessage.AlreadyDiscount);
+			} else if (generalDiscount?.discount?.code != code) {
+				await this.basketRepository.delete({ discountId: generalDiscount.discountId, userId });
+			}
 		}
 
-		await this.basketRepository.insert({ discountId: discount.id, userId });
+		await this.basketRepository.insert({ discountId: discount.id, userId, type: "total" });
 		return { message: BasketMessage.AddedDiscount };
 	}
